backend/pyproject.toml
[project]
name = "ats-backend"
version = "1.0.0"
description = "ATS Backend API for Shorthills AI"
requires-python = ">=3.12"

dependencies = [
    # Core framework
    "django>=5.0,<6.0",
    "djangorestframework>=3.15",

    # Auth
    "djangorestframework-simplejwt>=5.3",
    "social-auth-app-django>=5.4",

    # Database
    "psycopg[binary]>=3.1",
    "django-environ>=0.11",

    # API utilities
    "django-cors-headers>=4.3",
    "django-filter>=24.0",
    "drf-spectacular>=0.27",
    "django-ratelimit>=4.1",

    # Background tasks
    "celery>=5.3",
    "redis>=5.0",
    "django-celery-results>=2.5",
    "django-celery-beat>=2.6",

    # Resume parsing
    "pdfplumber>=0.11",
    "python-docx>=1.1",

    # AI
    "anthropic>=0.49",

    # Files & exports
    "openpyxl>=3.1",
    "django-storages>=1.14",
    "Pillow>=10.3",

    # Email / calendar
    "icalendar>=6.1",

    # Production server
    "gunicorn>=22.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.8",
    "factory-boy>=3.3",
    "pytest-cov>=5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["config", "apps"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.dev"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
